pyproject.toml
[tool.poetry]
name = "pptbench"
version = "0.1.1"
description = "A simple tool to benchmark the performance of PowerPoint layout."
authors = ["Tianyu Hu <tyrionhuu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pre-commit = "^4.0.1"
pipenv = "^2024.2.0"
flake8 = "^7.1.1"
matplotlib = "^3.9.3"
lxml = "^5.3.0"
python-pptx = "^1.0.2"
seaborn = "^0.13.2"
pandas = "^2.2.3"


[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.flake8]
ignore = "E203, E266, E501, W503"
max-line-length = 88
max-complexity = 18
select = "B,C,E,F,W,T4"

[tool.mypy]
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["./tests"]

# coverage report configuration
[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]

[tool.setuptools]
py-modules = ['pptbench']
